+++ assess-app/src/main/java/com/copower/pmcc/assess/dto/output/method/MdIncomeLeaseCostVo.java
@@ -2,6 +2,7 @@ package com.copower.pmcc.assess.dto.output.method;
 
 import com.copower.pmcc.assess.dal.basis.entity.MdIncomeLeaseCost;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -10,7 +11,7 @@ import java.util.Date;
 public class MdIncomeLeaseCostVo extends MdIncomeLeaseCost {
     private Date beginDate;
     private Date endDate;
-    private Integer yearCount;
+    private BigDecimal yearCount;
 
     public Date getBeginDate() {
         return beginDate;
@@ -28,11 +29,11 @@ public class MdIncomeLeaseCostVo extends MdIncomeLeaseCost {
         this.endDate = endDate;
     }
 
-    public Integer getYearCount() {
+    public BigDecimal getYearCount() {
         return yearCount;
     }
 
-    public void setYearCount(Integer yearCount) {
+    public void setYearCount(BigDecimal yearCount) {
         this.yearCount = yearCount;
     }
 }
